requirements.txt
google-generativeai>=0.8.0
flask>=3.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
redis>=5.0.0
PyJWT>=2.8.0
pyotp>=2.9.0
qrcode>=7.4.0
flasgger>=0.9.7
